`tests/wpt/web-platform-tests/html/browsers/origin/cross-origin-objects/cross-origin-objects.html`
was reverted to the upstream version.

+++ python/servo/util.py
@@ -9,21 +9,52 @@
 
 from __future__ import absolute_import, print_function, unicode_literals
 
+import hashlib
 import os
 import os.path
 import platform
 import shutil
 from socket import error as socket_error
-import StringIO
+import stat
+from io import BytesIO
 import sys
-import tarfile
+import time
 import zipfile
-import urllib2
+import six.moves.urllib as urllib
+
+
+try:
+    from ssl import HAS_SNI
+except ImportError:
+    HAS_SNI = False
+
+HAS_SNI_AND_RECENT_PYTHON = HAS_SNI and sys.version_info >= (2, 7, 9)
+
+
+def get_static_rust_lang_org_dist():
+    if HAS_SNI_AND_RECENT_PYTHON:
+        return "https://static.rust-lang.org/dist"
+
+    return "https://static-rust-lang-org.s3.amazonaws.com/dist"
+
+
+def get_urlopen_kwargs():
+    # The cafile parameter was added in 2.7.9
+    if HAS_SNI_AND_RECENT_PYTHON:
+        import certifi
+        return {"cafile": certifi.where()}
+    return {}
+
+
+def remove_readonly(func, path, _):
+    "Clear the readonly bit and reattempt the removal"
+    os.chmod(path, stat.S_IWRITE)
+    func(path)
 
 
 def delete(path):
     if os.path.isdir(path) and not os.path.islink(path):
-        shutil.rmtree(path)
+        shutil.rmtree(path, onerror=remove_readonly)
     else:
         os.remove(path)
 
@@ -37,11 +68,7 @@ def host_platform():
     elif os_type == "android":
         os_type = "linux-androideabi"
     elif os_type == "windows":
-        # If we are in a Visual Studio environment, use msvc
-        if os.getenv("PLATFORM") is not None:
-            os_type = "pc-windows-msvc"
-        else:
-            os_type = "unknown"
+        os_type = "pc-windows-msvc"
     elif os_type == "freebsd":
         os_type = "unknown-freebsd"
     else:
@@ -52,21 +79,14 @@ def host_platform():
 def host_triple():
     os_type = host_platform()
     cpu_type = platform.machine().lower()
-    if os_type.endswith("-msvc"):
-        # vcvars*.bat should set it properly
-        platform_env = os.environ.get("PLATFORM").upper()
-        if platform_env == "X86":
-            cpu_type = "i686"
-        elif platform_env == "X64":
-            cpu_type = "x86_64"
-        else:
-            cpu_type = "unknown"
-    elif cpu_type in ["i386", "i486", "i686", "i768", "x86"]:
+    if cpu_type in ["i386", "i486", "i686", "i768", "x86"]:
         cpu_type = "i686"
     elif cpu_type in ["x86_64", "x86-64", "x64", "amd64"]:
         cpu_type = "x86_64"
     elif cpu_type == "arm":
         cpu_type = "arm"
+    elif cpu_type == "aarch64":
+        cpu_type = "aarch64"
     else:
         cpu_type = "unknown"
 
@@ -75,24 +95,26 @@ def host_triple():
 
 def download(desc, src, writer, start_byte=0):
     if start_byte:
-        print("Resuming download of {}...".format(desc))
+        print("Resuming download of {} ...".format(src))
     else:
-        print("Downloading {}...".format(desc))
+        print("Downloading {} ...".format(src))
     dumb = (os.environ.get("TERM") == "dumb") or (not sys.stdout.isatty())
 
     try:
-        req = urllib2.Request(src)
+        req = urllib.request.Request(src)
         if start_byte:
-            req = urllib2.Request(src, headers={'Range': 'bytes={}-'.format(start_byte)})
-        resp = urllib2.urlopen(req)
+            req = urllib.request.Request(src, headers={'Range': 'bytes={}-'.format(start_byte)})
+        resp = urllib.request.urlopen(req, **get_urlopen_kwargs())
 
         fsize = None
-        if resp.info().getheader('Content-Length'):
-            fsize = int(resp.info().getheader('Content-Length').strip()) + start_byte
+        if resp.info().get('Content-Length'):
+            fsize = int(resp.info().get('Content-Length').strip()) + start_byte
 
         recved = start_byte
-        chunk_size = 8192
+        chunk_size = 64 * 1024
 
+        previous_progress_line = None
+        previous_progress_line_time = 0
         while True:
             chunk = resp.read(chunk_size)
             if not chunk:
@@ -101,23 +123,29 @@ def download(desc, src, writer, start_byte=0):
             if not dumb:
                 if fsize is not None:
                     pct = recved * 100.0 / fsize
-                    print("\rDownloading %s: %5.1f%%" % (desc, pct), end="")
+                    progress_line = "\rDownloading %s: %5.1f%%" % (desc, pct)
+                    now = time.time()
+                    duration = now - previous_progress_line_time
+                    if progress_line != previous_progress_line and duration > .1:
+                        print(progress_line, end="")
+                        previous_progress_line = progress_line
+                        previous_progress_line_time = now
 
                 sys.stdout.flush()
             writer.write(chunk)
 
         if not dumb:
             print()
-    except urllib2.HTTPError, e:
+    except urllib.error.HTTPError as e:
         print("Download failed ({}): {} - {}".format(e.code, e.reason, src))
         if e.code == 403:
             print("No Rust compiler binary available for this platform. "
                   "Please see https://github.com/servo/servo/#prerequisites")
         sys.exit(1)
-    except urllib2.URLError, e:
+    except urllib.error.URLError as e:
         print("Error downloading {}: {}. The failing URL was: {}".format(desc, e.reason, src))
         sys.exit(1)
-    except socket_error, e:
+    except socket_error as e:
         print("Looks like there's a connectivity issue, check your Internet connection. {}".format(e))
         sys.exit(1)
     except KeyboardInterrupt:
@@ -126,7 +154,7 @@ def download(desc, src, writer, start_byte=0):
 
 
 def download_bytes(desc, src):
-    content_writer = StringIO.StringIO()
+    content_writer = BytesIO()
     download(desc, src, content_writer)
     return content_writer.getvalue()
 
@@ -143,11 +171,41 @@ def download_file(desc, src, dst):
     os.rename(tmp_path, dst)
 
 
-def extract(src, dst, movedir=None):
-    if src.endswith(".zip"):
-        zipfile.ZipFile(src).extractall(dst)
-    else:
-        tarfile.open(src).extractall(dst)
+# https://stackoverflow.com/questions/39296101/python-zipfile-removes-execute-permissions-from-binaries
+# In particular, we want the executable bit for executable files.
+class ZipFileWithUnixPermissions(zipfile.ZipFile):
+    def extract(self, member, path=None, pwd=None):
+        if not isinstance(member, zipfile.ZipInfo):
+            member = self.getinfo(member)
+
+        if path is None:
+            path = os.getcwd()
+
+        extracted = self._extract_member(member, path, pwd)
+        mode = os.stat(extracted).st_mode
+        mode |= (member.external_attr >> 16)
+        os.chmod(extracted, mode)
+        return extracted
+
+    # For Python 3.x
+    def _extract_member(self, member, targetpath, pwd):
+        if sys.version_info[0] >= 3:
+            if not isinstance(member, zipfile.ZipInfo):
+                member = self.getinfo(member)
+
+            targetpath = super()._extract_member(member, targetpath, pwd)
+
+            attr = member.external_attr >> 16
+            if attr != 0:
+                os.chmod(targetpath, attr)
+            return targetpath
+        else:
+            return super(ZipFileWithUnixPermissions, self)._extract_member(member, targetpath, pwd)
+
+
+def extract(src, dst, movedir=None, remove=True):
+    assert src.endswith(".zip")
+    ZipFileWithUnixPermissions(src).extractall(dst)
 
     if movedir:
         for f in os.listdir(movedir):
@@ -156,4 +214,18 @@ def extract(src, dst, movedir=None):
             os.rename(frm, to)
         os.rmdir(movedir)
 
-    os.remove(src)
+    if remove:
+        os.remove(src)
+
+
+def check_hash(filename, expected, algorithm):
+    hasher = hashlib.new(algorithm)
+    with open(filename, "rb") as f:
+        while True:
+            block = f.read(16 * 1024)
+            if len(block) == 0:
+                break
+            hasher.update(block)
+    if hasher.hexdigest() != expected:
+        print("Incorrect {} hash for {}".format(algorithm, filename))
+        sys.exit(1)
